requirements.txt
django
djangorestframework
# WSGI server for production
gunicorn
PyJWT
psycopg2-binary
pgvector
google-generativeai
openai
django-environ
psycopg2-binary
